+++ src/components/sidebar/routePathView/routePathListTab/RoutePathListNode.tsx
@@ -15,7 +15,7 @@ import SubSites from '~/routing/subSites';
 import { CodeListStore } from '~/stores/codeListStore';
 import { RoutePathStore } from '~/stores/routePathStore';
 import NodeUtils from '~/utils/NodeUtils';
-import TransitTypeHelper from '~/utils/TransitTypeHelper';
+import TransitTypeUtils from '~/utils/TransitTypeUtils';
 import InputContainer from '../../../controls/InputContainer';
 import TextContainer from '../../../controls/TextContainer';
 import RoutePathListItem from './RoutePathListItem';
@@ -282,7 +282,7 @@ class RoutePathListNode extends React.Component<IRoutePathListNodeProps> {
             icon = this.addBorder(icon, '#353333');
         } else if (node.type === NodeType.STOP) {
             node.transitTypes!.forEach(type => {
-                icon = this.addBorder(icon, TransitTypeHelper.getColor(type));
+                icon = this.addBorder(icon, TransitTypeUtils.getColor(type));
             });
         }
